requirements.txt
streamlit==1.32.0
requests==2.31.0
pandas==2.2.1
python-dotenv==1.0.1
chardet